Add separate case for asset path '.'
In a debug build where '.' is one of the asset paths, all local asset
names previously had to start with './' in order to be accessible, while
a production build did not show this behaviour.

This commit handles '.' separately in debug builds, so that
`getAsset("path/to/file")` works as well as `getAsset("./path/to/file")`.